feat: implement AnalysisLayout and refactor analysis page to utilize new layout component
