+++ frontend/src/app/(public)/analysis/[id]/page.tsx
@@ -1,8 +1,25 @@
 import { Grid } from '@mui/material';
 import { notFound } from 'next/navigation';
+import { cache } from 'react';
 import { AnalyticsCard } from '@/components/features/AnalyticsCard';
 import { TranscriptionCard } from '@/components/features/TranscriptionCard';
-import { getAnalysisData } from './layout'; // ✅ Importa a função do novo layout
+import { AnalysisLayout } from '@/components/layouts/AnalysisLayout'; // ✅ Importa o novo componente de layout
+
+// --- Mock de Dados ---
+const getAnalysisData = cache(async (id: string) => {
+  console.log(`Buscando dados para o ID: ${id}`);
+  if (id === 'not-found') return null;
+  return {
+    id,
+    fileName: 'audio_file_name.mp3',
+    transcription: 'Mussum Ipsum, cacilds vidis litro abertis...',
+    analytics: [
+      { id: 'metric1', title: 'Análise de Sentimento' },
+      { id: 'metric2', title: 'Densidade de Palavras-chave' },
+      { id: 'metric3', title: 'Tópicos Principais' },
+    ],
+  };
+});
 
 interface AnalysisPageProps {
   params: { id: string };
@@ -10,31 +27,36 @@ interface AnalysisPageProps {
 }
 
 export default async function AnalysisPage({ params, searchParams }: AnalysisPageProps) {
-  const currentView = searchParams.view || 'transcription';
-
   const data = await getAnalysisData(params.id);
-
+  
   if (!data) {
     notFound();
   }
 
-  if (currentView === 'analytics') {
-    return (
-      <Grid container spacing={4}>
-        {data.analytics.map((metric) => (
-          <Grid size={{xs: 12, lg: 6, xl: 4}} key={metric.id}>
-            <AnalyticsCard title={metric.title}>
-              <div className="bg-gray-200 h-64 flex items-center justify-center rounded-lg">
-                Gráfico para {metric.title}
-              </div>
-            </AnalyticsCard>
-          </Grid>
-        ))}
-      </Grid>
-    );
-  }
+  const currentView = searchParams.view || 'transcription';
 
   return (
-    <TranscriptionCard transcriptionText={data.transcription} />
+    // ✅ A página agora se envolve com o AnalysisLayout
+    <AnalysisLayout data={data}>
+      {/*
+        O conteúdo abaixo será passado como `children` para o AnalysisLayout
+        e renderizado dentro da tag <main> dele.
+      */}
+      {currentView === 'analytics' ? (
+        <Grid container spacing={4}>
+          {data.analytics.map((metric) => (
+            <Grid size={{xs: 12, lg: 6, xl: 4}} key={metric.id}>
+              <AnalyticsCard title={metric.title}>
+                <div className="bg-gray-200 h-64 flex items-center justify-center rounded-lg">
+                  Gráfico para {metric.title}
+                </div>
+              </AnalyticsCard>
+            </Grid>
+          ))}
+        </Grid>
+      ) : (
+        <TranscriptionCard transcriptionText={data.transcription} />
+      )}
+    </AnalysisLayout>
   );
 }
