+++ frontend/src/app/layout.tsx
@@ -41,7 +41,7 @@ export default function RootLayout({
             <CssBaseline />
             <div className='flex flex-col min-h-screen'>
               <Header/>
-              <main className='flex-grow'>
+              <main className='flex-grow flex flex-col'>
                 {children}
               </main>
               <Footer/>
