+++ frontend/src/components/layouts/AnalysisSideBar.tsx
@@ -1,44 +1,47 @@
 'use client';
 
-import { Box, Divider, Drawer, List, ListItem, ListItemButton, ListItemIcon, ListItemText, Toolbar, Typography } from '@mui/material';
+import { Box, Divider, Drawer, List, ListItem, ListItemButton, ListItemIcon, ListItemText, Typography } from '@mui/material';
 import { FileAudio, FileText, LayoutGrid } from 'lucide-react';
 import Link from 'next/link';
 import { useSearchParams } from 'next/navigation';
-import { Logo } from './Logo';
 
 const DRAWER_WIDTH = 280;
 
 interface AnalysisSidebarProps {
   fileName: string;
-  basePath: string; // ex: /analysis/123
+  basePath: string;
 }
 
 export function AnalysisSidebar({ fileName, basePath }: AnalysisSidebarProps) {
   const searchParams = useSearchParams();
-  const currentView = searchParams.get('view') || 'transcription'; 
+  const currentView = searchParams.get('view') || 'transcription';
 
   const navItems = [
     { text: 'Transcrição', icon: <FileText />, view: 'transcription' },
     { text: 'Análise', icon: <LayoutGrid />, view: 'analytics' },
   ];
 
-  const drawerContent = (
-    <div>
-      <Toolbar sx={{ justifyContent: 'center', py: 2 }}>
-        <Logo/>
-      </Toolbar>
-
-      <Divider />
-     
+  return (
+    <Drawer
+      variant="permanent"
+      sx={{
+        width: DRAWER_WIDTH,
+        flexShrink: 0,
+        ['& .MuiDrawer-paper']: {
+          width: DRAWER_WIDTH,
+          boxSizing: 'border-box',
+          position: 'relative', 
+          borderRight: 'none',
+        },
+      }}
+    >
       <Box sx={{ p: 2, display: 'flex', alignItems: 'center', gap: 1.5 }}>
         <FileAudio className="text-gray-500" />
         <Typography variant="body2" color="text.secondary" noWrap>
           {fileName}
         </Typography>
       </Box>
-
       <Divider />
-
       <List>
         {navItems.map((item) => (
           <ListItem key={item.text} disablePadding>
@@ -63,23 +66,6 @@ export function AnalysisSidebar({ fileName, basePath }: AnalysisSidebarProps) {
           </ListItem>
         ))}
       </List>
-    </div>
-  );
-
-  return (
-    <Drawer
-      variant="permanent"
-      sx={{
-        width: DRAWER_WIDTH,
-        flexShrink: 0,
-        ['& .MuiDrawer-paper']: { 
-          width: DRAWER_WIDTH, 
-          boxSizing: 'border-box',
-          borderRight: '1px solid rgba(0, 0, 0, 0.12)'
-        },
-      }}
-    >
-      {drawerContent}
     </Drawer>
   );
 }
